Cleaned some things up in shared-tests module with subpackages
